Organize ipc_bytes
